test function to create and delete test dir works
