+++ test/create.test.js
@@ -2,25 +2,19 @@ const assert = require('assert');
 const fs = require('fs');
 const run = require('./run.js');
 const simpleDB = require('../lib/create.js');
-const dataObj = require('../lib/data.js');
 
 describe('create operations for database', function() {
-    it('create database', function() {
-        const testDir = 'db-test-dir';
-        const db = simpleDB.createDB(testDir);
+    const testDir = 'db-test-dir';
 
-        before(done => {
-            fs.rmdirSync(testDir);
-            db();
-        });
-
-
+    before(() => {
+        if (fs.existsSync('./' + testDir)) {
+            fs.rmdirSync('./' + testDir);
+        }
+        simpleDB.createDB(testDir);
     });
 
-    it('create data object for storing in database', function() {
-        const data = { name: 'steve', color: 'blue' };
-        const output = dataObj.createDataObj(data);
-        assert.equal(output.name, 'steve');
-        assert.equal(output.color, 'blue');
+    it('create database', function() {
+
+
     });
 });
